Fix /workers and /workers/<int:id> routes
